move content of feec/multipatch/api to api/feec and api/discretization
